Currently a new variable for the input spectrum is used, called "spectrum" (line 21). So the ArgumentParser is deactivated.
An example spectrum is uploaded to run pahfit: 'NGC7023-NW-BRIGHT.txt'

A drude profile function is implemented (line 83).

You can chose between drude and lorentzian profile. This is stored in the "fitprof" variable (line 34)

The pop-up plot is deactivated since it only shows the current best fit every "x" steps of iterations and shows no model components. In addition when the fit stops the fit in the window is not the best fit.

Plotting of all components and best fit is implemented.

Risiduals and best model are calculated.

The fit results are shown as a multi-line print instead of a 1-line print, and are saved into a file.

The saved fit resutls are then used to read and construct the multi-component plot.