Handle wave_bounds that differ from instrumet segment coverage range
When a user passes a spectrum and specifies an instrument segment or
set of segments, it is possible the user-provided wavelength coverage
differs from the specified segment range, due to lost or trimmed data,
etc.  This handles this case, as well as wavelength ranges which fall
"off the end" of the coverage, warning for <3% spill-over, and raising
an error for larger amounts.