Simplify feature bounds using np.nan instead of masking
Partially masked arrays are not well supported, and required elaborate
workaround for pretty printing.  Here we use a structured array type,
and represent "fixed" bounds with np.nan instead of masking.  I.e. non
is interpreted as the same as negative or positive infinity, unless
both bound are nan, which indicates a fixed parameter.