Array vs scalar arguments to determine fitted/fixed in Fitter
Clean up Features contents before giving them to Fitter. Nan is
converted to np.inf or -np.inf. Fixed values (both bounds nan) are
converted to a single scalar. The feature adding functions of Fitter
will interpret scalars as fixed parameters, and 3-arrays as the initial
value and bounds of parameters that will be fit.